streamlit_dashboard/requirements.txt
streamlit
pandas
kafka-python
plotly